diplom/requirements.txt
amqp==5.1.1
anyio==3.7.0
asgiref==3.6.0
attrs==23.1.0
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
dadata==21.10.1
defusedxml==0.7.1
Django==4.2.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
google-auth==2.17.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.13.1
keyrings.google-artifactregistry-auth==1.1.2
kombu==5.2.4
MarkupSafe==2.1.2
more-itertools==9.1.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2
drf-extra-fields==3.5.0